+++ src/template/index.js
@@ -1,6 +1,8 @@
 'use strict';
 
 module.exports = {
+  adminLoginHTML: require('./admin-login-html'),
+  adminIAHTML: require('./admin-ia-html'),
   adminOverviewHTML: require('./admin-overview-html'),
   adminTopicDetailsHTML: require('./admin-topic-details-html'),
   badgeSVG: require('./badge-svg'),